Added ability to display custom view controller as the content view
