+++ TabbedSplitViewController/TabbedSplitViewController/TabbedSplitViewController.swift
@@ -26,7 +26,7 @@ public struct PKTabBarItem<T> {
     public var navigationBarSelectedImage: UIImage?
     /// Declares that the tab opens a screen that takes all the space
     ///   beside the tab bar that otherwise would show master and detail screens.
-    public let isFullScreen: Bool
+    public let isFullWidth: Bool
     /// An action value that will be passed to the `OnSelection` callback
     ///
     /// Default types are `UIViewController` for the main tab bar and
@@ -34,7 +34,7 @@ public struct PKTabBarItem<T> {
     public let action: T
 
     ///
-    public init(title: String, image: UIImage, selectedImage: UIImage? = nil, navigationBarImage: UIImage? = nil, navigationBarSelectedImage: UIImage? = nil, isFullScreen: Bool = false, action: T) {
+    public init(title: String, image: UIImage, selectedImage: UIImage? = nil, navigationBarImage: UIImage? = nil, navigationBarSelectedImage: UIImage? = nil, isFullWidth: Bool = false, action: T) {
         self.title = title
         self.image = image
         self.selectedImage = selectedImage
@@ -42,7 +42,7 @@ public struct PKTabBarItem<T> {
 
         self.navigationBarImage = navigationBarImage
         self.navigationBarSelectedImage = navigationBarSelectedImage
-        self.isFullScreen = isFullScreen
+        self.isFullWidth = isFullWidth
     }
 }
 
@@ -103,7 +103,7 @@ public class TabbedSplitViewController: UIViewController {
                 || detailViewMinWidth != oldValue.detailViewMinWidth
         }
 
-        fileprivate static let zero: Configuration = Configuration(tabBarWidth: 0, masterViewWidth: 0, detailViewMinWidth: 0, tabBarBackgroundColor: .white, detailBackgroundColor: .white, verticalSeparatorColor: .gray, detailAsModalShouldStayInPlace: false, showTabBarAsSideBarWithSizeChange: nil, showMasterAsSideBarWithSizeChange: nil, showDetailAsModalWithSizeChange: nil)
+        static let zero: Configuration = Configuration(tabBarWidth: 0, masterViewWidth: 0, detailViewMinWidth: 0, tabBarBackgroundColor: .white, detailBackgroundColor: .white, verticalSeparatorColor: .gray, detailAsModalShouldStayInPlace: false, showTabBarAsSideBarWithSizeChange: nil, showMasterAsSideBarWithSizeChange: nil, showDetailAsModalWithSizeChange: nil)
     }
 
     /// Tabbed Split View Controller Configuration
@@ -153,24 +153,22 @@ public class TabbedSplitViewController: UIViewController {
     ///   before a detail screen opened and after a detail screen closed.
     public var defaultDetailViewController: UIViewController? {
         didSet {
-            detailVC.defaultViewController = self.defaultDetailViewController
+            masterDetailVC.defaultDetailViewController = self.defaultDetailViewController
         }
     }
 
-    private let masterVC = PKMasterViewController()
-    private let detailVC = PKDetailViewController()
     private let tabBarVC = PKTabBar()
 
     private let mainView: PKTabbedSplitView
-    private let masterDetailView: MasterDetailContentView
+    private let masterDetailVC: MasterDetailContentViewController
+    private var currentContentVC: UIViewController
 
     private var configured: Bool = false
 
     private var futureTraits: UITraitCollection?
     private var futureSize: CGSize?
 
-    private var sideBarView: UIView?
-    private var sideNavigationBarViewController: UIViewController?
+    private var sideBarViewController: UIViewController?
 
     private var state: State = (true, true, true)
 
@@ -178,22 +176,27 @@ public class TabbedSplitViewController: UIViewController {
 
     public init(items: [PKTabBarItem<UIViewController>], actionItems: [PKTabBarItem<TabBarAction>] = [], config: Configuration? = nil) {
         self.config = config ?? Configuration()
-        masterDetailView = MasterDetailContentView(masterView: masterVC.view, detailView: detailVC.view)
-        mainView = PKTabbedSplitView(tabBarView: tabBarVC.view, contentView: masterDetailView)
+        masterDetailVC = MasterDetailContentViewController()
+        mainView = PKTabbedSplitView(tabBarView: tabBarVC.view, contentView: masterDetailVC.view)
+        currentContentVC = masterDetailVC
 
         super.init(nibName: nil, bundle: nil)
+
 
         update(oldConfig: .zero)
         tabBarVC.tabBar.items = items
         tabBarVC.actionsBar.items = actionItems
+
     }
 
     public required init?(coder aDecoder: NSCoder) {
         self.config = Configuration()
-        masterDetailView = MasterDetailContentView(masterView: masterVC.view, detailView: detailVC.view)
-        mainView = PKTabbedSplitView(tabBarView: tabBarVC.view, contentView: masterDetailView)
+        masterDetailVC = MasterDetailContentViewController()
+        mainView = PKTabbedSplitView(tabBarView: tabBarVC.view, contentView: masterDetailVC.view)
+        currentContentVC = masterDetailVC
 
         super.init(coder: aDecoder)
+
 
         update(oldConfig: .zero)
     }
@@ -208,23 +211,35 @@ public class TabbedSplitViewController: UIViewController {
         super.viewDidLoad()
 
         tabBarVC.tabBar.didSelectCallback = { [unowned self] item, selectedIndex in
-            let isTheSameItem = (item.action == self.masterVC.viewController)
-            if !isTheSameItem {
+            let isTheSameMasterVC = item.action == self.masterDetailVC.masterViewController
+            let itemChanged = (item.isFullWidth == (self.currentContentVC == self.masterDetailVC))
+                                || !isTheSameMasterVC
+
+            if itemChanged {
                 self.selectedTabBarItemIndex = selectedIndex
-                self.masterVC.viewController = item.action
+                if item.isFullWidth {
+                    self.replaceContent(with: item.action)
+                    return
+                }
+
+                // Master-detail part until the end of the closure
+
+                self.resetContentToMasterDetail()
+                self.masterDetailVC.masterViewController = item.action
 
                 self.logger?.log("Hide tab bar: \(self.mainView.hideTabBarView)")
-                self.logger?.log("Hide master view: \(self.masterDetailView.hideMasterView)")
+                self.logger?.log("Hide master view: \(self.masterDetailVC.hideMasterView)")
             }
+
             if self.mainView.hideTabBarView {
                 // Hide navigation view while opening a detail
                 self.mainView.hideSideBar()
             }
-            else if self.masterDetailView.hideMasterView {
+            else if self.masterDetailVC.hideMasterView {
                 if self.mainView.sideBarIsHidden {
                     self.mainView.showSideBar()
                     self.tabBarVC.tabBar.isOpen = true
-                } else if isTheSameItem {
+                } else if isTheSameMasterVC {
                     self.mainView.hideSideBar()
                     self.tabBarVC.tabBar.isOpen = false
                 } else {
@@ -241,8 +256,7 @@ public class TabbedSplitViewController: UIViewController {
         }
 
         addChild(tabBarVC)
-        addChild(masterVC)
-        addChild(detailVC)
+        addChild(masterDetailVC)
 
         update(oldConfig: .zero)
 
@@ -276,35 +290,32 @@ public class TabbedSplitViewController: UIViewController {
 
         if let hideMaster = config.showMasterAsSideBarWithSizeChange?(screenSize, traits, config) {
             // Update only if it's changed
-            if masterDetailView.hideMasterView != hideMaster {
-                masterDetailView.hideMasterView = hideMaster
-                if hideMaster {
-                    let masterSideBar = masterDetailView.prepareMasterForSideBar()
-                    mainView.addSideBar(masterSideBar)
-                    self.sideBarView = masterSideBar
+            if self.masterDetailVC.hideMasterView != hideMaster {
+                self.masterDetailVC.hideMasterView = hideMaster
+                if hideMaster, let sideBar = self.masterDetailVC.prepareMasterForSideBar() {
+                    mainView.addSideBar(sideBar.view)
+                    self.sideBarViewController = sideBar
                 }
             }
             tabBarVC.tabBar.shouldDisplayArrow = hideMaster
             state.masterHidden = hideMaster
         }
-        masterVC.didMove(toParent: self)
 
         // Hide detail from main view if there is not enough width
         if let hideDetail = config.showDetailAsModalWithSizeChange?(screenSize, traits, config) {
-            if masterDetailView.hideDetailView != hideDetail {
-                masterDetailView.hideDetailView = hideDetail
+            if self.masterDetailVC.hideDetailView != hideDetail {
+                self.masterDetailVC.hideDetailView = hideDetail
                 if hideDetail {
-                    masterDetailView.removeDetailView(removeFromViewHierarchy: !config.detailAsModalShouldStayInPlace)
+                    self.masterDetailVC.removeDetailView(removeFromViewHierarchy: !config.detailAsModalShouldStayInPlace)
                 }
             }
             state.detailHidden = hideDetail
         }
-        detailVC.didMove(toParent: self)
         self.state = state
 
         tabBarVC.tabBarWidthConstraint.isActive = true
         tabBarVC.actionBarWidthConstraint.isActive = true
-        masterDetailView.masterViewWidthConstraint.isActive = true
+        self.masterDetailVC.masterViewWidthConstraint.isActive = true
 
         super.viewWillAppear(animated)
         configured = true
@@ -349,8 +360,8 @@ public class TabbedSplitViewController: UIViewController {
                 hideTabBar = hideTabBarFunc(realSize, traits, self.config)
                 self.logger?.log("hide Tab Bar: \(hideTabBar)")
             }
-            updateDetail = self.masterDetailView.hideDetailView != hideDetail
-            updateMaster = self.masterDetailView.hideMasterView != hideMaster
+            updateDetail = self.masterDetailVC.hideDetailView != hideDetail
+            updateMaster = self.masterDetailVC.hideMasterView != hideMaster
             updateTabBar = self.mainView.hideTabBarView != hideTabBar
 
             self.state = (hideTabBar, hideMaster, hideDetail)
@@ -358,7 +369,7 @@ public class TabbedSplitViewController: UIViewController {
             guard updateDetail || updateMaster || updateTabBar else { return }
 
             if updateMaster, !hideMaster {
-                self.masterDetailView.hideMasterView = false
+                self.masterDetailVC.hideMasterView = false
             }
             if updateDetail, !hideDetail, !self.config.detailAsModalShouldStayInPlace {
                 self.hideDetailAsModal()
@@ -373,36 +384,37 @@ public class TabbedSplitViewController: UIViewController {
                 self.removeNavigationSideBar(keepTabBarHidden: keepTabBarHidden)
             }
             if updateMaster, !hideMaster {
-                if let view = self.sideBarView {
-                    self.sideBarView = nil
-                    self.mainView.removeSideBar(view)
-                    self.masterDetailView.putMasterBack()
+                if let sideBar = self.sideBarViewController {
+                    self.sideBarViewController = nil
+                    self.mainView.removeSideBar(sideBar.view)
+                    self.masterDetailVC.putMasterBack()
                 }
             }
             if updateDetail, !hideDetail {
                 if self.config.detailAsModalShouldStayInPlace {
                     self.hideDetailAsModalInPlace()
                 }
-                self.masterDetailView.addDetailView()
+                self.masterDetailVC.addDetailView()
             }
             // Then, removing from the stack view
             if updateTabBar, hideTabBar {
                 self.addNavigationSideBar()
             }
             if updateMaster, hideMaster, !hideDetail {
-                let sideBarView = self.masterDetailView.prepareMasterForSideBar()
-                self.mainView.addSideBar(sideBarView)
-                self.sideBarView = sideBarView
+                if let sideBar = self.masterDetailVC.prepareMasterForSideBar() {
+                    self.mainView.addSideBar(sideBar.view)
+                    self.sideBarViewController = sideBar
+                }
             }
         }, completion: { _ in
             if updateMaster, hideMaster, !hideDetail {
-                self.masterDetailView.hideMasterView = true
+                self.masterDetailVC.hideMasterView = true
             }
             if updateTabBar {
                 self.mainView.hideTabBarView = hideTabBar
             }
             if updateDetail {
-                self.masterDetailView.hideDetailView = hideDetail
+                self.masterDetailVC.hideDetailView = hideDetail
                 if hideDetail {
                     self.presentDetailAsModal()
                 }
@@ -419,13 +431,13 @@ public class TabbedSplitViewController: UIViewController {
             if detailViewController != nil {
                 presentDetailInPlace()
             } else {
-                self.masterDetailView.removeDetailView(removeFromViewHierarchy: true)
+                self.masterDetailVC.removeDetailView(removeFromViewHierarchy: true)
             }
         } else {
-            self.masterDetailView.removeDetailView(removeFromViewHierarchy: true)
+            self.masterDetailVC.removeDetailView(removeFromViewHierarchy: true)
             if let detail = detailViewController {
                 // Remove the view controller from the DetailVC, but keep it saved in TSVC
-                detailVC.setViewController(nil, animate: false)
+                masterDetailVC.setDetailViewController(nil, animate: false)
                 detail.view.translatesAutoresizingMaskIntoConstraints = true
                 self.present(detail, animated: false)
             }
@@ -434,7 +446,7 @@ public class TabbedSplitViewController: UIViewController {
     private func hideDetailAsModal() {
         if let detail = detailViewController {
             dismiss(animated: false) {
-                self.detailVC.setViewController(detail, animate: false)
+                self.masterDetailVC.setDetailViewController(detail, animate: false)
             }
         }
     }
@@ -449,7 +461,7 @@ public class TabbedSplitViewController: UIViewController {
         if !state.tabBarHidden {
             mainView.hideTabBar(animator: animator)
         }
-        masterDetailView.presentFullWidthDetailView(animator: animator, hideMaster: !state.masterHidden)
+        self.masterDetailVC.presentFullWidthDetailView(animator: animator, hideMaster: !state.masterHidden)
         mainView.setSideBarGestureRecognizerEnabled(false)
 
         animator.startAnimation()
@@ -466,7 +478,7 @@ public class TabbedSplitViewController: UIViewController {
             mainView.showTabBar(animator: animator)
         }
 
-        masterDetailView.closeFullWidthDetailView(animator: animator,
+        self.masterDetailVC.closeFullWidthDetailView(animator: animator,
                                                   keepShown: keepShown,
                                                   showMaster: !state.masterHidden)
         mainView.setSideBarGestureRecognizerEnabled(true)
@@ -476,19 +488,41 @@ public class TabbedSplitViewController: UIViewController {
 
     private func addNavigationSideBar() {
         let navVC = configureNavigationBar(tabBarVC.tabBar.items, tabBarVC.actionsBar.items, tabBarVC.tabBar.didSelectCallback!, tabBarVC.actionsBar.didSelectCallback!, selectedTabBarItemIndex)
-        sideNavigationBarViewController = navVC
+        sideBarViewController = navVC
         addChild(navVC)
         mainView.addNavigationBar(navVC.view)
         navVC.didMove(toParent: self)
     }
     private func removeNavigationSideBar(keepTabBarHidden: Bool) {
-        guard let navVC = sideNavigationBarViewController else { return }
+        guard let navVC = sideBarViewController else { return }
 
         navVC.willMove(toParent: nil)
         self.mainView.removeSideBar(navVC.view)
         self.mainView.putTabBarBack(keepHidden: keepTabBarHidden)
         navVC.removeFromParent()
-        sideNavigationBarViewController = nil
+        sideBarViewController = nil
+    }
+
+    private func replaceContent(with contentVC: UIViewController) {
+        guard contentVC != self.currentContentVC else {
+            return
+        }
+
+        let prevVC = self.currentContentVC
+        self.currentContentVC = contentVC
+
+        self.addChild(contentVC)
+        contentVC.willMove(toParent: self)
+        self.mainView.replaceContentView(with: contentVC.view) {
+            prevVC.removeFromParent()
+            prevVC.viewIfLoaded?.removeFromSuperview()
+            contentVC.didMove(toParent: self)
+        }
+
+
+    }
+    private func resetContentToMasterDetail() {
+        self.replaceContent(with: self.masterDetailVC)
     }
 
     // MARK: - Public functions
@@ -499,31 +533,31 @@ public class TabbedSplitViewController: UIViewController {
     }
     public func showDetailViewController(_ vc: UIViewController, completion: (() -> Void)? = nil) {
         // Show Detail screen if needed
-        if masterDetailView.hideDetailView {
+        if self.masterDetailVC.hideDetailView {
             // Hide master view while opening a detail
             mainView.hideSideBar()
 
             if config.detailAsModalShouldStayInPlace {
                 presentDetailInPlace()
-                detailVC.setViewController(vc, animate: false, completion: completion)
+                masterDetailVC.setDetailViewController(vc, animate: false, completion: completion)
             } else {
                 present(vc, animated: true, completion: completion)
             }
         } else {
-            if masterDetailView.hideMasterView {
+            if self.masterDetailVC.hideMasterView {
                 // Hide master view while opening a detail
                 mainView.hideSideBar()
             }
-            detailVC.setViewController(vc, animate: true, completion: completion)
+            masterDetailVC.setDetailViewController(vc, animate: true, completion: completion)
         }
         self.detailViewController = vc
     }
 
     public func dismissDetailViewController(animated flag: Bool = true, completion: (() -> Void)? = nil) {
-        if masterDetailView.hideDetailView {
+        if self.masterDetailVC.hideDetailView {
             if config.detailAsModalShouldStayInPlace {
                 hideDetailInPlace(keepShown: false, then: {
-                    self.detailVC.setViewController(nil, animate: false, completion: completion)
+                    self.masterDetailVC.setDetailViewController(nil, animate: false, completion: completion)
                 })
             } else {
                 dismiss(animated: flag, completion: completion)
@@ -531,7 +565,7 @@ public class TabbedSplitViewController: UIViewController {
         }
         else if detailViewController != nil {
             logger?.log("Removing presented detail VC from parent VC")
-            detailVC.setViewController(nil, animate: true, completion: completion)
+            masterDetailVC.setDetailViewController(nil, animate: true, completion: completion)
         }
         self.detailViewController = nil
     }
@@ -577,161 +611,13 @@ public class TabbedSplitViewController: UIViewController {
         if config.tabBarWidth != oldConfig.tabBarWidth {
             tabBarVC.tabBarWidth = config.tabBarWidth
         }
-        if config.masterViewWidth != oldConfig.masterViewWidth {
-            masterDetailView.masterViewWidth = config.masterViewWidth
-        }
         if config.tabBarBackgroundColor != oldConfig.tabBarBackgroundColor {
             tabBarVC.backgroundColor = config.tabBarBackgroundColor
         }
-        if config.detailBackgroundColor != oldConfig.detailBackgroundColor {
-            detailVC.backgroundColor = config.detailBackgroundColor
-        }
         if config.verticalSeparatorColor != oldConfig.verticalSeparatorColor {
             tabBarVC.verticalSeparatorColor = config.verticalSeparatorColor
-            masterVC.verticalSeparatorColor = config.verticalSeparatorColor
         }
-    }
-
-}
-
-// MARK: - Master view controller
-
-private class PKMasterViewController: UIViewController {
-
-    fileprivate var viewController: UIViewController? {
-        didSet {
-            if let prev = oldValue {
-                prev.willMove(toParent: nil)
-                prev.view.removeFromSuperview()
-                prev.removeFromParent()
-            }
-            if let next = viewController {
-                addChild(next)
-                addChildView(next.view)
-                if shouldAddVerticalSeparator {
-                    view.addVerticalSeparator(verticalSeparator, color: verticalSeparatorColor)
-                }
-                view.layoutIfNeeded()
-                next.didMove(toParent: self)
-            }
-        }
-    }
-    fileprivate var shouldAddVerticalSeparator: Bool = true
-    fileprivate var verticalSeparatorColor: UIColor = .gray {
-        didSet {
-            verticalSeparator.backgroundColor = verticalSeparatorColor
-        }
-    }
-    private let verticalSeparator = VerticalSeparatorView()
-
-    fileprivate init() {
-        super.init(nibName: nil, bundle: nil)
-    }
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    fileprivate override func viewDidLoad() {
-        super.viewDidLoad()
-
-        if shouldAddVerticalSeparator {
-            view.addVerticalSeparator(verticalSeparator, color: verticalSeparatorColor)
-        }
-
-        view.accessibilityIdentifier = "Master View"
-        view.backgroundColor = .white
-    }
-
-}
-
-// MARK: - Detail view controller
-
-private class PKDetailViewController: UIViewController {
-
-    private var viewController: UIViewController?
-
-    fileprivate var defaultViewController: UIViewController? {
-        didSet {
-            // Don't replace current VC if it's presented
-            if viewController == nil, defaultViewController != nil {
-                setViewController(defaultViewController, animate: false)
-            }
-        }
-    }
-    fileprivate var backgroundColor: UIColor = .white {
-        didSet {
-            view.backgroundColor = backgroundColor
-        }
-    }
-
-    fileprivate init() {
-        super.init(nibName: nil, bundle: nil)
-    }
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    fileprivate override func viewDidLoad() {
-        super.viewDidLoad()
-
-        view.accessibilityIdentifier = "Detail View"
-        view.backgroundColor = backgroundColor
-    }
-
-    fileprivate func setViewController(_ newVC: UIViewController?, animate: Bool, completion: (() -> Void)? = nil) {
-        let oldVC = viewController
-        viewController = newVC ?? defaultViewController
-        replaceViewController(oldVC, with: viewController, animate: animate, completion: completion)
-    }
-
-    private func replaceViewController(_ oldVC: UIViewController?, with newVC: UIViewController?, animate: Bool, completion: (() -> Void)?) {
-
-        if let next = newVC {
-            addChild(next)
-            addChildViewCentered(next.view)
-        }
-        oldVC?.willMove(toParent: nil)
-
-        let completion = { [unowned self] in
-            oldVC?.removeFromParent()
-            oldVC?.view.removeFromSuperview()
-            newVC?.didMove(toParent: self)
-            completion?()
-        }
-
-        guard animate else {
-            completion()
-            return
-        }
-
-        if let prev = oldVC, let next = newVC {
-            // Switching between two view controllers
-            if let nextXPosition = next.view.constraint(for: .centerX) {
-                nextXPosition.constant = -view.bounds.width
-                view.layoutIfNeeded()
-
-                nextXPosition.constant = 0
-                prev.view.constraint(for: .centerX)?.constant = view.bounds.width
-            }
-
-            transition(from: prev, to: next, duration: 0.33, options: [.curveEaseInOut], animations: {
-                self.view.layoutIfNeeded()
-            }) { finished in
-                if finished {
-                    completion()
-                }
-            }
-        } else if let view = oldVC?.view ?? newVC?.view {
-            let isClosing = (newVC == nil)
-            view.alpha = isClosing ? 1.0 : 0.0
-            UIView.transition(with: view, duration: 0.33, options: [.curveEaseInOut], animations: {
-                view.alpha = isClosing ? 0.0 : 1.0
-            }) { finished in
-                if finished {
-                    completion()
-                }
-            }
-        }
+        masterDetailVC.config = config
     }
 
 }
